[SCRIPTS] Check files if exist

+++ scripts/setup_venv.sh
@@ -5,9 +5,18 @@ if [ -z "$VIRTUAL_ENV" ]; then
 else
     python -m pip install --upgrade pip wheel setuptools
 
-    python -m pip install --upgrade -r requirements-versioned.txt
-    python -m pip install --upgrade -r requirements-git.txt
-    python -m pip install --upgrade -r requirements-dev.txt
+    if [ -e requirements-versioned.txt ]; then
+        python -m pip install --upgrade -r requirements-versioned.txt
+    fi
+    if [ -e requirements-git.txt ]; then
+        python -m pip install --upgrade -r requirements-git.txt
+    fi
+    if [ -e requirements-dev.txt ]; then
+        python -m pip install --upgrade -r requirements-dev.txt
+    fi
+    if [ -e requirements.txt ]; then
+        python -m pip install --upgrade -r requirements.txt
+    fi
 
     python -m pip install --upgrade -e .
